+++ jest.config.js
@@ -1,7 +1,0 @@
-/** @type {import('@ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'jsdom',
-  collectCoverage: true,
-  coverageDirectory: 'coverage',
-}